Renamed learningPanel to learningOverlay
